requirements.txt
alembic==1.8.1
appnope==0.1.3
APScheduler==3.6.3
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
aspose-cells==22.8.0
async-generator==1.10
attrs==22.1.0
backcall==0.2.0
backports.zoneinfo==0.2.1
beautifulsoup4==4.11.1
bleach==5.0.1
boto==2.49.0
Bottleneck==1.3.2
brotlipy==0.7.0
cachetools==4.2.2
certifi==2020.6.20
certipy==0.1.3
cffi==1.14.3
chardet==3.0.4
click==8.1.3
cloudpickle==1.6.0
clyent==1.2.2
colorama==0.4.4
conda==4.13.0
conda-build==3.20.5
conda-package-handling==1.7.2
conda-verify==3.4.2
constants==0.6.0
contextlib2==0.6.0.post1
cryptography==3.1.1
cycler==0.10.0
Cython==0.29.21
cytoolz==0.11.0
dask==2.30.0
decorator==4.4.2
defusedxml==0.6.0
demes==0.2.2
diff-match-patch==20200713
distlib==0.3.5
distributed==2.30.1
dj-database-url==0.5.0
docutils==0.16
entrypoints==0.3
et-xmlfile==1.0.1
ez-setup==0.9
fastcache==1.1.0
fastjsonschema==2.16.1
filelock==3.7.1
flake8==3.8.4
Flask==2.1.3
Flask-Cors==3.0.10
Flask-RESTful==0.3.9
fsspec==0.8.3
future==0.18.2
gevent==20.9.0
glob2==0.7
gmpy2==2.0.8
google-api-core==2.8.2
google-api-python-client==2.55.0
google-auth==2.9.1
google-auth-httplib2==0.1.0
google-auth-oauthlib==0.5.2
googleapis-common-protos==1.56.4
greenlet==0.4.17
gunicorn==20.1.0
h5py==2.10.0
HeapDict==1.0.1
html5lib==1.1
httplib2==0.20.4
humanize==4.2.3
idna==2.10
imageio==2.9.0
imagesize==1.2.0
importlib-metadata==4.12.0
importlib-resources==5.9.0
inflection==0.5.1
iniconfig==1.1.1
intervaltree==3.1.0
ipykernel==5.3.4
ipython==7.19.0
ipython_genutils==0.2.0
ipywidgets==7.5.1
isort==5.6.4
itsdangerous==2.1.2
jdcal==1.4.1
jedi==0.17.1
Jinja2==3.1.2
joblib==0.17.0
JPype1==1.4.0
json5==0.9.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==6.1.7
jupyter-console==6.2.0
jupyter-core==4.11.1
jupyter-server==1.18.1
jupyter-telemetry==0.1.0
jupyterhub==2.3.1
jupyterlab==2.2.6
jupyterlab-pygments==0.1.2
jupyterlab-server==2.15.0
kastore==0.3.2
keyring==21.4.0
kiwisolver==1.3.0
lazy-object-proxy==1.4.3
libarchive-c==2.9
llvmlite==0.34.0
locket==0.2.0
logging==0.4.9.6
lxml==4.6.1
Mako==1.2.1
MarkupSafe==2.1.1
matplotlib==3.3.2
mccabe==0.6.1
mistune==0.8.4
mkl-fft==1.2.0
mkl-random==1.1.1
mkl-service==2.3.0
mock==4.0.2
more-itertools==8.6.0
mpmath==1.1.0
msgpack==1.0.0
msprime==1.2.0
multipledispatch==0.6.0
Nasdaq-Data-Link==1.0.2
navigator-updater==0.2.1
nbclient==0.5.1
nbconvert==6.5.0
nbformat==5.4.0
nest-asyncio==1.4.2
networkx==2.5
newick==1.3.2
nltk==3.5
nose==1.3.7
notebook==6.1.4
numba==0.51.2
numexpr==2.7.1
numpy==1.19.2
numpydoc==1.1.0
oauthlib==3.2.0
olefile==0.46
openpyxl==3.0.5
packaging==20.4
pamela==1.0.0
pandas==1.1.3
pandocfilters==1.4.3
parso==0.7.0
partd==1.1.0
path==15.0.0
pathlib2==2.3.5
pathtools==0.1.2
patsy==0.5.1
pep8==1.7.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==8.0.1
pip==22.2.2
pipenv==2022.7.24
pkginfo==1.6.1
platformdirs==2.5.2
pluggy==0.13.1
ply==3.11
prometheus-client==0.8.0
prompt-toolkit==3.0.8
protobuf==4.21.4
psutil==5.7.2
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycodestyle==2.6.0
pycosat==0.6.3
pycparser==2.20
pycurl==7.43.0.6
pydocstyle==5.1.1
pyflakes==2.2.0
Pygments==2.7.2
pylint==2.6.0
pyodbc==4.0.0-unsupported
pyOpenSSL==19.1.0
pyparsing==2.4.7
pyrsistent==0.17.3
pyslim==0.700
PySocks==1.7.1
pytest==0.0.0
python-dateutil==2.8.1
python-decouple==3.6
python-json-logger==2.0.4
python-jsonrpc-server==0.4.0
python-language-server==0.35.1
python-telegram-bot==13.13
pytz==2020.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.1.1
PyYAML==5.3.1
pyzmq==19.0.2
QDarkStyle==2.8.1
QtAwesome==1.0.1
qtconsole==4.7.7
QtPy==1.9.0
Quandl==3.7.0
regex==2020.10.15
requests==2.24.0
requests-oauthlib==1.3.1
rope==0.18.0
rsa==4.9
Rtree==0.9.4
ruamel_yaml==0.15.87
scikit-image==0.17.2
scikit-learn==0.23.2
scipy==1.5.2
seaborn==0.11.0
Send2Trash==1.5.0
setuptools==50.3.1.post20201107
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.15.0
sniffio==1.2.0
snowballstemmer==2.0.0
sortedcollections==1.2.1
sortedcontainers==2.2.2
soupsieve==2.0.1
Sphinx==3.2.1
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sphinxcontrib-websupport==1.2.4
spyder==4.1.5
spyder-kernels==1.9.4
SQLAlchemy==1.3.20
statsmodels==0.12.0
svgwrite==1.4.3
sympy==1.6.2
tables==3.6.1
tblib==1.7.0
terminado==0.9.1
testpath==0.4.4
threadpoolctl==2.1.0
tifffile==2020.10.1
tinycss2==1.1.1
toml==0.10.1
toolz==0.11.1
tornado==6.2
tqdm==4.50.2
traitlets==5.3.0
tskit==0.5.2
typing-extensions==3.7.4.3
tzdata==2022.1
tzlocal==4.2
ujson==4.0.1
unicodecsv==0.14.1
uritemplate==4.1.1
urllib3==1.25.11
virtualenv==20.16.1
virtualenv-clone==0.5.7
voila==0.3.6
watchdog==0.10.3
wcwidth==0.2.5
webencodings==0.5.1
websocket-client==1.3.3
websockets==10.3
Werkzeug==2.2.1
wheel==0.35.1
whitenoise==6.2.0
widgetsnbextension==3.5.1
workbook==1.1
wrapt==1.11.2
wurlitzer==2.0.1
xlrd==1.2.0
XlsxWriter==1.3.7
xlutils==2.0.0
xlwings==0.20.8
xlwt==1.3.0
xmltodict==0.12.0
yapf==0.30.0
zict==2.0.0
zipp==3.4.0
zope.event==4.5.0
zope.interface==5.1.2
